hypervisor: kvm: RMI: Initialize vCPU
KVM doesn't allow modifying PSTATE at runtime, since that register is
private to the Realm. Skip PSTATE when RMI is enabled.

This is incomplete, as the set/get_state functions also access PSTATE.
However getting/setting the state of a running confidential VM is not
supported, at least not at a register level. Migration would require
special support from the Realm hypervisor, since guest registers aren't
visible from the host. KVM returns an error when trying to access the
confidential guest state so that's what we get when trying to
save/restore the VM, and that's seems reasonable for now.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>