vmm: Force DMA accesses with arm_rmi
Enable "force_iommu" when arm_rmi is enabled. It's a misnomer because
there isn't necessarily an IOMMU involved, it just tells the guest to
perform DMA on virtio as on a normal device - explicitly share memory
with the device so that the host is able to access it. It's needed for
arm_rmi because by default all memory is private to the guest.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>